[artf155298] - Updates on portal, mapping email, emailConfig, notification to the context not more to user.

+++ simple2secure.portal/src/main/java/com/simple2secure/portal/controller/SettingsController.java
@@ -9,6 +9,7 @@
 package com.simple2secure.portal.controller;
 
 import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -32,81 +33,83 @@ import com.simple2secure.portal.repository.SettingsRepository;
 import com.simple2secure.portal.service.MessageByLocaleService;
 
 @RestController
+@RequestMapping("/api/settings/")
 public class SettingsController {
-	
+
 	@Autowired
 	SettingsRepository settingsRepository;
-	
+
 	@Autowired
 	LicensePlanRepository licensePlanRepository;
-	
+
 	@Autowired
 	ConfigRepository configRepository;
-		
-    @Autowired
-    MessageByLocaleService messageByLocaleService;
+
+	@Autowired
+	MessageByLocaleService messageByLocaleService;
 
 	@SuppressWarnings({ "rawtypes", "unchecked" })
-	@RequestMapping(value = "/api/settings", method = RequestMethod.GET)
+	@RequestMapping(value = "", method = RequestMethod.GET)
 	@PreAuthorize("hasAuthority('SUPERADMIN')")
 	public ResponseEntity<SettingsDTO> getSettings(@RequestHeader("Accept-Language") String locale) {
 		List<Settings> settings = settingsRepository.findAll();
 		List<LicensePlan> licensePlans = licensePlanRepository.findAll();
-		if(settings != null) {
-			if(settings.size() == 1) {
+		if (settings != null) {
+			if (settings.size() == 1) {
 				return new ResponseEntity<SettingsDTO>(new SettingsDTO(settings.get(0), licensePlans), HttpStatus.OK);
 			}
-			else {
-				return new ResponseEntity(new CustomErrorType(messageByLocaleService.getMessage("problem_occured_while_retrieving_settings", locale)), HttpStatus.NOT_FOUND);
-			}			
 		}
-		else {
-			return new ResponseEntity(new CustomErrorType(messageByLocaleService.getMessage("problem_occured_while_retrieving_settings", locale)), HttpStatus.NOT_FOUND);
-		}
+		return new ResponseEntity(new CustomErrorType(messageByLocaleService.getMessage("problem_occured_while_retrieving_settings", locale)),
+				HttpStatus.NOT_FOUND);
+
 	}
-	
+
 	@SuppressWarnings({ "unchecked", "rawtypes" })
-	@RequestMapping(value = "/api/settings", method = RequestMethod.POST, consumes = "application/json")
+	@RequestMapping(value = "", method = RequestMethod.POST, consumes = "application/json")
 	@PreAuthorize("hasAuthority('SUPERADMIN')")
-	public ResponseEntity<Settings> updateSettings(@RequestBody Settings settings, @RequestHeader("Accept-Language") String locale) throws ItemNotFoundRepositoryException {
-		if(settings != null) {
+	public ResponseEntity<Settings> updateSettings(@RequestBody Settings settings, @RequestHeader("Accept-Language") String locale)
+			throws ItemNotFoundRepositoryException {
+		if (settings != null) {
 			settingsRepository.update(settings);
 			return new ResponseEntity<Settings>(settings, HttpStatus.OK);
 		}
-		else {
-			return new ResponseEntity(new CustomErrorType(messageByLocaleService.getMessage("problem_occured_while_updating_settings", locale)), HttpStatus.NOT_FOUND);
-		}
+		return new ResponseEntity(new CustomErrorType(messageByLocaleService.getMessage("problem_occured_while_updating_settings", locale)),
+				HttpStatus.NOT_FOUND);
 	}
-	
+
 	@SuppressWarnings({ "unchecked", "rawtypes" })
-	@RequestMapping(value = "/api/settings/licensePlan", method = RequestMethod.POST, consumes = "application/json")
+	@RequestMapping(value = "/licensePlan", method = RequestMethod.POST, consumes = "application/json")
 	@PreAuthorize("hasAuthority('SUPERADMIN')")
-	public ResponseEntity<LicensePlan> saveLicensePlan(@RequestBody LicensePlan licensePlan, @RequestHeader("Accept-Language") String locale) throws ItemNotFoundRepositoryException {
-		if(licensePlan != null) {
-			if(Strings.isNullOrEmpty(licensePlan.getId())) {
+	public ResponseEntity<LicensePlan> saveLicensePlan(@RequestBody LicensePlan licensePlan, @RequestHeader("Accept-Language") String locale)
+			throws ItemNotFoundRepositoryException {
+		if (licensePlan != null) {
+			if (Strings.isNullOrEmpty(licensePlan.getId())) {
 				licensePlanRepository.save(licensePlan);
-			}
-			else {
+			} else {
 				licensePlanRepository.update(licensePlan);
 			}
 			return new ResponseEntity<LicensePlan>(licensePlan, HttpStatus.OK);
 		}
-		else {
-			return new ResponseEntity(new CustomErrorType(messageByLocaleService.getMessage("problem_occured_while_updating_settings", locale)), HttpStatus.NOT_FOUND);
-		}
-	}	
-	
+		return new ResponseEntity(new CustomErrorType(messageByLocaleService.getMessage("problem_occured_while_updating_settings", locale)),
+				HttpStatus.NOT_FOUND);
+	}
+
 	@SuppressWarnings({ "unchecked", "rawtypes" })
-	@RequestMapping(value = "/api/settings/licensePlan/{licensePlanId}", method = RequestMethod.DELETE)
+	@RequestMapping(value = "/licensePlan/{licensePlanId}", method = RequestMethod.DELETE)
 	@PreAuthorize("hasAuthority('SUPERADMIN')")
-	public ResponseEntity<LicensePlan> deleteLicensePlan(@PathVariable("licensePlanId") String licensePlanId, @RequestHeader("Accept-Language") String locale) throws ItemNotFoundRepositoryException {
-		LicensePlan licensePlan = licensePlanRepository.find(licensePlanId);
-		if(licensePlan != null) {
-			licensePlanRepository.delete(licensePlan);
-			return new ResponseEntity<LicensePlan>(licensePlan, HttpStatus.OK);
+	public ResponseEntity<LicensePlan> deleteLicensePlan(@PathVariable("licensePlanId") String licensePlanId,
+			@RequestHeader("Accept-Language") String locale) throws ItemNotFoundRepositoryException {
+
+		if (!Strings.isNullOrEmpty(licensePlanId)) {
+			LicensePlan licensePlan = licensePlanRepository.find(licensePlanId);
+			if (licensePlan != null) {
+				licensePlanRepository.delete(licensePlan);
+				return new ResponseEntity<LicensePlan>(licensePlan, HttpStatus.OK);
+			}
 		}
-		else {
-			return new ResponseEntity(new CustomErrorType(messageByLocaleService.getMessage("problem_occured_while_deleting_license_plan", locale)), HttpStatus.NOT_FOUND);
-		}
+
+		return new ResponseEntity(new CustomErrorType(messageByLocaleService.getMessage("problem_occured_while_deleting_license_plan", locale)),
+				HttpStatus.NOT_FOUND);
+
 	}
 }
